requirements.txt
anyio==4.1.0
appdirs==1.4.4
asttokens==2.4.1
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.2.0
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.46.0
frozendict==2.3.10
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
ipykernel==6.27.1
ipython==8.12.3
jedi==0.19.1
jupyter-client==8.6.0
jupyter-core==5.5.0
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
multitasking==0.0.11
nest-asyncio==1.5.8
nixtlats==0.1.18
numpy==1.24.4
packaging==23.2
pandas==2.0.3
pandas-datareader==0.10.0
parso==0.8.3
peewee==3.17.0
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.13
pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
seaborn==0.13.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
tenacity==8.2.3
tornado==6.4
traitlets==5.14.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
utilsforecast==0.0.19
wcwidth==0.2.12
webencodings==0.5.1
yfinance==0.2.32
zipp==3.17.0
